Split `filteredTransactions` into filtered & sorted
Split the `filteredTransactions` state variable into two separate variables: `filteredTransactions` and sortedTransactions. In the `ControlPanel` (which manages the displayed data), `filteredTransactions` is now only updated when any filtering changes occur, while `sortedTransactions` is always updated (note that this variable contains both filtered and sorted data). The `Balance` component was refactored to use `filteredTransactions`. This change ensures that the chart only re-renders when there are changes to the transaction data, not when items simply change positions within the array;